+++ HW2/.pytest_cache/.gitignore
@@ -1,0 +1,2 @@
+# Created by pytest automatically.
+v
