+++ HW2_2/requirements22.txt
@@ -1,0 +1,1 @@
+pytest
