--- HW32/test_hw2.py
@@ -1,15 +1,0 @@
-from hw2 import cache
-
-
-def test_cache():
-    def func(a, b):
-        return (a ** b) ** 2
-
-    cache_func = cache(func)
-
-    some = 100, 200
-
-    val_1 = cache_func(*some)
-    val_2 = cache_func(*some)
-
-    assert val_1 is val_2
